Performance fixes / code improvements
Turns out `Aliases.at()` was super slow --especially since we called
it so much when resolving aliases during indexing. This was due to it
having to re-parse the document multiple times for each alias
reference. Moving to sourceror and allowing the AST to be sent in cut
the indexing time for lexical and all its deps from 60 seconds to 10.