First pass at concurrent indexer
Prior, we were looking at all the modules in a project and indexing
them, which was somewhat problematic, as we were gobbling up erlang
modules as well as elixir ones.

To start simpler, this just looks at your project's .ex and .exs
files (including deps) and indexes them. It might be a bit
overzealous, as it'll pick up _anything_ in the project root, but this
can be tamed by using Mix and asking it where your sources and deps
live and just indexing those files.

TODO: We likely need to index both elixir and erlang's standard
library, but this is a bit fraught, as it's not always available.